+++ full-moon/tests/roblox_cases/pass/variadic_type_packs/source.lua
@@ -1,7 +1,0 @@
-function foo<T...>(param: () -> T...)
-end
-
-type Foo<T...> = () -> T...
-
-function bar<T...>(...: T...)
-end